(PC-29962)[PRO] test: Move isSameOffers tests to their own file in the utils folder.
