+++ pro/src/pages/Offers/utils/__specs__/isSameOffer.spec.ts
@@ -1,7 +1,7 @@
 import { collectiveOfferFactory } from 'utils/collectiveApiFactories'
 import { listOffersOfferFactory } from 'utils/individualApiFactories'
 
-import { isSameOffer } from '../utils'
+import { isSameOffer } from '../isSameOffer'
 
 describe('isSameOffer', () => {
   const individualOffer = listOffersOfferFactory()
