+++ pro/src/screens/Offers/Offers.tsx
@@ -34,7 +34,7 @@ import strokeUserIcon from 'icons/stroke-user.svg'
 import { Offers as OffersContainer } from 'pages/Offers/Offers/Offers'
 import { CollectiveOffersActionsBar } from 'pages/Offers/Offers/OffersActionsBar/CollectiveOffersActionsBar'
 import { IndividualOffersActionsBar } from 'pages/Offers/Offers/OffersActionsBar/IndividualOffersActionsBar'
-import { isSameOffer } from 'pages/Offers/Offers/utils'
+import { isSameOffer } from 'pages/Offers/utils/isSameOffer'
 import { selectCurrentOffererId } from 'store/user/selectors'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
 import { ButtonVariant } from 'ui-kit/Button/types'
